only create vm if creating

+++ web3/apps/vms/forms.py
@@ -35,17 +35,19 @@ class VirtualMachineForm(forms.ModelForm):
         hostname = slugify(instance.name).replace("_", "-")
 
         if commit:
+            editing = bool(instance.pk)
             instance.save()
             self.save_m2m()
-            ret = call_api("container.create", name=hostname, template=self.cleaned_data.get("template", "debian"))
-            if ret is None or ret[0] == 1:
-                client.captureMessage("Failed to create VM: {}".format(ret))
-                instance.delete()
-                return None
-            else:
-                if ret[0] != 2:
-                    instance.uuid = uuid.UUID(ret[1].split("\n")[-1])
-                    instance.save()
+            if editing:
+                ret = call_api("container.create", name=hostname, template=self.cleaned_data.get("template", "debian"))
+                if ret is None or ret[0] == 1:
+                    client.captureMessage("Failed to create VM: {}".format(ret))
+                    instance.delete()
+                    return None
+                else:
+                    if ret[0] != 2:
+                        instance.uuid = uuid.UUID(ret[1].split("\n")[-1])
+                        instance.save()
 
         return instance
